requirements.txt
dnspython
colorama
win_unicode_console
